# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Personal site API — blog posts, notes, projects, and about page
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Post {
  id: number;
  slug: string;
  title: string;
  content: string;
  excerpt: string;
  published: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface PostInput {
  /** @minLength 1 */
  slug: string;
  /** @minLength 1 */
  title: string;
  content: string;
  excerpt: string;
  published: boolean;
}

export interface Note {
  id: number;
  slug: string;
  title: string;
  content: string;
  excerpt: string;
  published: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface NoteInput {
  /** @minLength 1 */
  slug: string;
  /** @minLength 1 */
  title: string;
  content: string;
  excerpt: string;
  published: boolean;
}

export interface Project {
  id: number;
  title: string;
  description: string;
  year: number;
  articleUrl?: string | null;
  demoUrl?: string | null;
  sourceUrl?: string | null;
  published: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface ProjectInput {
  /** @minLength 1 */
  title: string;
  description: string;
  year: number;
  articleUrl?: string | null;
  demoUrl?: string | null;
  sourceUrl?: string | null;
  published: boolean;
}

export interface About {
  name: string;
  tagline: string;
  intro: string;
  bio: string;
  emailNewsletter: string;
  blueskyUrl: string;
  rssUrl: string;
  updatedAt: string;
}

export interface AboutInput {
  name: string;
  tagline: string;
  intro: string;
  bio: string;
  emailNewsletter: string;
  blueskyUrl: string;
  rssUrl: string;
}

export interface DashboardSummary {
  postsTotal: number;
  postsPublished: number;
  notesTotal: number;
  notesPublished: number;
  projectsTotal: number;
  projectsPublished: number;
}

export type ActivityItemKind =
  (typeof ActivityItemKind)[keyof typeof ActivityItemKind];

export const ActivityItemKind = {
  post: "post",
  note: "note",
  project: "project",
} as const;

export interface ActivityItem {
  kind: ActivityItemKind;
  id: number;
  title: string;
  slug?: string | null;
  published?: boolean;
  updatedAt: string;
}

export type ListPostsParams = {
  /**
   * Filter by published status. Omit to include all (admin).
   */
  published?: boolean;
};

export type ListNotesParams = {
  published?: boolean;
};

export type ListProjectsParams = {
  published?: boolean;
};