# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Personal site API — blog posts, notes, projects, and about page
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: posts
    description: Blog posts
  - name: notes
    description: Personal notes
  - name: projects
    description: Personal projects
  - name: about
    description: About me content
  - name: dashboard
    description: Dashboard summary endpoints
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /posts:
    get:
      operationId: listPosts
      tags: [posts]
      summary: List blog posts
      parameters:
        - in: query
          name: published
          schema:
            type: boolean
          description: Filter by published status. Omit to include all (admin).
      responses:
        "200":
          description: List of posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Post"
    post:
      operationId: createPost
      tags: [posts]
      summary: Create a blog post (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostInput"
      responses:
        "201":
          description: Post created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
  /posts/{slug}:
    get:
      operationId: getPost
      tags: [posts]
      summary: Get post by slug
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "404":
          description: Not found
  /posts/id/{id}:
    put:
      operationId: updatePost
      tags: [posts]
      summary: Update a post (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostInput"
      responses:
        "200":
          description: Updated post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
    delete:
      operationId: deletePost
      tags: [posts]
      summary: Delete a post (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /notes:
    get:
      operationId: listNotes
      tags: [notes]
      summary: List notes
      parameters:
        - in: query
          name: published
          schema:
            type: boolean
      responses:
        "200":
          description: List of notes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Note"
    post:
      operationId: createNote
      tags: [notes]
      summary: Create a note (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NoteInput"
      responses:
        "201":
          description: Note created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
  /notes/{slug}:
    get:
      operationId: getNote
      tags: [notes]
      summary: Get note by slug
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
        "404":
          description: Not found
  /notes/id/{id}:
    put:
      operationId: updateNote
      tags: [notes]
      summary: Update a note (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NoteInput"
      responses:
        "200":
          description: Updated note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
    delete:
      operationId: deleteNote
      tags: [notes]
      summary: Delete a note (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List projects
      parameters:
        - in: query
          name: published
          schema:
            type: boolean
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a project (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Project created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}:
    put:
      operationId: updateProject
      tags: [projects]
      summary: Update a project (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /about:
    get:
      operationId: getAbout
      tags: [about]
      summary: Get about page content
      responses:
        "200":
          description: About content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/About"
    put:
      operationId: updateAbout
      tags: [about]
      summary: Update about page content (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AboutInput"
      responses:
        "200":
          description: Updated about content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/About"

  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Counts of posts, notes, projects (admin)
      responses:
        "200":
          description: Summary counts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
  /dashboard/recent:
    get:
      operationId: getRecentActivity
      tags: [dashboard]
      summary: Recent activity feed across content (admin)
      responses:
        "200":
          description: Recent items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityItem"

components:
  schemas:
    HealthStatus:
      type: object
      required: [status]
      properties:
        status:
          type: string
    Post:
      type: object
      required: [id, slug, title, content, excerpt, published, createdAt, updatedAt]
      properties:
        id:
          type: integer
        slug:
          type: string
        title:
          type: string
        content:
          type: string
        excerpt:
          type: string
        published:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    PostInput:
      type: object
      required: [slug, title, content, excerpt, published]
      properties:
        slug:
          type: string
          minLength: 1
        title:
          type: string
          minLength: 1
        content:
          type: string
        excerpt:
          type: string
        published:
          type: boolean
    Note:
      type: object
      required: [id, slug, title, content, excerpt, published, createdAt, updatedAt]
      properties:
        id:
          type: integer
        slug:
          type: string
        title:
          type: string
        content:
          type: string
        excerpt:
          type: string
        published:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    NoteInput:
      type: object
      required: [slug, title, content, excerpt, published]
      properties:
        slug:
          type: string
          minLength: 1
        title:
          type: string
          minLength: 1
        content:
          type: string
        excerpt:
          type: string
        published:
          type: boolean
    Project:
      type: object
      required: [id, title, description, year, published, createdAt, updatedAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        year:
          type: integer
        articleUrl:
          type: string
          nullable: true
        demoUrl:
          type: string
          nullable: true
        sourceUrl:
          type: string
          nullable: true
        published:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ProjectInput:
      type: object
      required: [title, description, year, published]
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        year:
          type: integer
        articleUrl:
          type: string
          nullable: true
        demoUrl:
          type: string
          nullable: true
        sourceUrl:
          type: string
          nullable: true
        published:
          type: boolean
    About:
      type: object
      required: [name, tagline, intro, bio, emailNewsletter, blueskyUrl, rssUrl, updatedAt]
      properties:
        name:
          type: string
        tagline:
          type: string
        intro:
          type: string
        bio:
          type: string
        emailNewsletter:
          type: string
        blueskyUrl:
          type: string
        rssUrl:
          type: string
        updatedAt:
          type: string
          format: date-time
    AboutInput:
      type: object
      required: [name, tagline, intro, bio, emailNewsletter, blueskyUrl, rssUrl]
      properties:
        name:
          type: string
        tagline:
          type: string
        intro:
          type: string
        bio:
          type: string
        emailNewsletter:
          type: string
        blueskyUrl:
          type: string
        rssUrl:
          type: string
    DashboardSummary:
      type: object
      required: [postsTotal, postsPublished, notesTotal, notesPublished, projectsTotal, projectsPublished]
      properties:
        postsTotal:
          type: integer
        postsPublished:
          type: integer
        notesTotal:
          type: integer
        notesPublished:
          type: integer
        projectsTotal:
          type: integer
        projectsPublished:
          type: integer
    ActivityItem:
      type: object
      required: [kind, id, title, updatedAt]
      properties:
        kind:
          type: string
          enum: [post, note, project]
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
          nullable: true
        published:
          type: boolean
        updatedAt:
          type: string
          format: date-time