# lib/api-zod/src/generated/types/about.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Personal site API — blog posts, notes, projects, and about page
 * OpenAPI spec version: 0.1.0
 */

export interface About {
  name: string;
  tagline: string;
  intro: string;
  bio: string;
  emailNewsletter: string;
  blueskyUrl: string;
  rssUrl: string;
  updatedAt: Date;
}