# lib/api-zod/src/generated/types/activityItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Personal site API — blog posts, notes, projects, and about page
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityItemKind } from "./activityItemKind";

export interface ActivityItem {
  kind: ActivityItemKind;
  id: number;
  title: string;
  slug?: string | null;
  published?: boolean;
  updatedAt: Date;
}