# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Personal site API — blog posts, notes, projects, and about page
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  postsTotal: number;
  postsPublished: number;
  notesTotal: number;
  notesPublished: number;
  projectsTotal: number;
  projectsPublished: number;
}