# lib/api-zod/src/generated/types/noteInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Personal site API — blog posts, notes, projects, and about page
 * OpenAPI spec version: 0.1.0
 */

export interface NoteInput {
  /** @minLength 1 */
  slug: string;
  /** @minLength 1 */
  title: string;
  content: string;
  excerpt: string;
  published: boolean;
}