# lib/api-zod/src/generated/types/projectInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Personal site API — blog posts, notes, projects, and about page
 * OpenAPI spec version: 0.1.0
 */

export interface ProjectInput {
  /** @minLength 1 */
  title: string;
  description: string;
  year: number;
  articleUrl?: string | null;
  demoUrl?: string | null;
  sourceUrl?: string | null;
  published: boolean;
}